+++ converters/terminal/src/syntax.rs
@@ -161,6 +161,7 @@ mod tests {
             index_entries: std::rc::Rc::new(std::cell::RefCell::new(Vec::new())),
             has_valid_index_section: false,
             list_indent: std::rc::Rc::new(std::cell::Cell::new(0)),
+            warnings: acdc_converters_core::WarningSink::default(),
         }
     }